👑Another fix

+++ commands/admin/deleteChannel.ts
@@ -15,6 +15,8 @@ const CreateChannelCommand: Command = {
     async run(client, message, args) {
         const channel = message.mentions.channels.first() || message.guild?.channels.cache.get(args[0])
 
+        if (!channel) return message.channel.send("**Please mention a channel to delete**")
+
         // @ts-ignore
         message.guild?.channels.cache.get(channel?.id)?.delete()
